split views.py into different views
